\name{RcppTN-package}
\alias{RcppTN-package}
\alias{RcppTN}
\docType{package}
\title{
Truncated Normal Distribution RNG
}
\description{
This package provides an R function to generate random deviates
from a Truncated Normal distribution using the Robert (1995)
algorithm. No d/p/q functions are provided. See the 'truncnorm'
package for these facilities.
}
\details{
\tabular{ll}{
Package: \tab RcppTN\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-04-11\cr
License: \tab GPL (>=2)\cr
}


}

\author{
Jonathan Olmsted

Maintainer: Jonathan Olmsted <jpolmsted@gmail.com>
}

\references{

  Robert, Christian P. "Simulation of truncated normal
  variables". Statistics and Computing 5.2 (1995):
  121-125. <http://dx.doi.org/10.1007/BF00143942>

}
\keyword{ package }
\keyword{ distribution }